+++ ui/temboardui/plugins/dashboard/static/js/temboard.dashboard.js
@@ -2,24 +2,13 @@
 $(function() {
   "use strict";
 
-  function html_error_modal(code, error) {
-    var html = '<div class="modal" id="ErrorModal" tabindex="-1" role="dialog" aria-labelledby="ErrorModalLabel" aria-hidden="true">';
-    html += '   <div class="modal-dialog">';
-    html += '     <div class="modal-content">';
-    html += '       <div class="modal-header">';
-    html += '         <h4 class="modal-title" id="ErrorModalLabel">Error '+code+'</h4>';
-    html += '         <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>';
-    html += '       </div>';
-    html += '       <div class="modal-body">';
-    html += '         <div class="alert alert-danger" role="alert">'+error+'</div>';
-    html += '       </div>';
-    html += '       <div class="modal-footer" id="ErrorModalFooter">';
-    html += '         <button type="button" class="btn btn-outline-secondary" data-dismiss="modal">Close</button>';
-    html += '       </div>';
-    html += '     </div>';
-    html += '   </div>';
-    html += '</div>';
-    return html;
+  function html_error(code, error) {
+    return `
+    <div class="alert alert-danger" role="alert">
+      <h4 class="modal-title" id="ErrorModalLabel">Error ${code}</h4>
+      <p>${error}</p>
+    </div>
+    `;
   }
 
   /*
@@ -27,19 +16,22 @@ $(function() {
    * updateDashboard() callback.
    */
   function refreshDashboard() {
+    var start_time = $('#pg_uptime time').attr("datetime");
+    $('#pg_uptime time').text((moment(start_time).fromNow()));
+    $('#pg_uptime time').attr("title", (moment(start_time).format("LLLL")));
+
     $.ajax({
       url: '/proxy/'+agent_address+'/'+agent_port+'/dashboard',
       type: 'GET',
       async: true,
       contentType: "application/json",
       success: function (data) {
-        $('#ErrorModal').modal('hide');
+        $('#divError').html('');
         updateDashboard(data, true);
         updateTps([data]);
         updateLoadaverage([data]);
       },
       error: function(xhr) {
-        $('#ErrorModal').modal('hide');
         if (xhr.status == 401 || xhr.status == 302) {
           // force a reload of the page, should lead to the server login page
           location.href = location.href;
@@ -51,8 +43,7 @@ $(function() {
         } else {
           code = '';
         }
-        $('#modalError').html(html_error_modal(code, error));
-        $('#ErrorModal').modal('show');
+        $('#divError').html(html_error(code, error));
       }
     });
   }
@@ -69,7 +60,9 @@ $(function() {
     $('#nb_db').html(databases ? databases.databases : null);
     $('#pg_data').html(data['pg_data']);
     $('#pg_port').html(data['pg_port']);
-    $('#pg_uptime').html(data['pg_uptime'] || 'N/A');
+    if (!'pg_start_time' in data) {
+      $('#pg_uptime').html(data['pg_uptime'] || 'N/A');  // For 7.X
+    }
 
     /** Update memory usage chart **/
     memorychart.data.datasets[0].data[0] = data['memory']['active'];